Refactor location handling and update UI.
This commit refactors location management, simplifies UI components, and modernizes the codebase.

- Replaced the lifecycle-aware `LocationManager` with a more integrated approach in `BaseSunInfoActivity` using `requestLocationUpdates` directly with a `ViewModel`.
- Removed the `SwipeRefreshLayout` from the main screen as location updates are now continuous.
- Replaced the text selection dialog in `LiberActivity` with a `Spinner` for a more modern and direct user experience.
- Renamed `options_menu.xml` to `actions_menu.xml` and updated menu item icons to Material Symbols for a consistent design.
- Eliminated several utility classes (`LocationUtil`, `ToastUtil`, `TypedArrayAccessor`) by inlining their functionality, reducing boilerplate and improving code clarity.
- Standardized view binding usage across activities for consistency.

+++ helios/src/main/java/org/onereed/helios/sun/SunAzimuthInfo.java
@@ -23,7 +23,7 @@ public abstract class SunAzimuthInfo {
     double azimuthNow = parameters.on(when).execute().getAzimuth();
     Instant when1 = when.plus(DELTA_TIME);
     double azimuthSoon = parameters.on(when1).execute().getAzimuth();
-    double deltaAzimuth = DirectionUtil.zeroCenterDeg(azimuthSoon - azimuthNow);
+    float deltaAzimuth = DirectionUtil.zeroCenterDeg(azimuthSoon - azimuthNow);
 
     return create((float) azimuthNow, /* clockwise= */ deltaAzimuth >= 0.0);
   }
